+++ src/frontend/src/forms/PurchaseOrderForms.tsx
@@ -220,7 +220,9 @@ export function usePurchaseOrderLineItemFields({
     }
 
     if (create) {
-      fields['merge_items'] = {};
+      fields['merge_items'] = {
+        default: globalSettings.isSet('PURCHASEORDER_MERGE_LINE_ITEMS', true)
+      };
     }
 
     return fields;